+++ tests/testthat/test_classoutput.R
@@ -78,12 +78,12 @@ df_patient01_tobra <- data.frame(ID=1,TIME=c(0.0,1.0,14.0),
                                  EVID=c(1,0,0),
                                  CLCREAT=80,WT=65)
 
-patient01_tobra      <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
-                          dat=df_patient01_tobra)
-
 test_that("poso_simu_pop returns the expected objects", {
-  p01_pop_mod    <- poso_simu_pop(patient01_tobra)
-  p01_pop_nomod  <- poso_simu_pop(patient01_tobra,return_model = FALSE)
+  p01_pop_mod    <- poso_simu_pop(dat=df_patient01_tobra,
+                                  prior_model=mod_tobramycin_2cpt_fictional)
+  p01_pop_nomod  <- poso_simu_pop(dat=df_patient01_tobra,
+                                  prior_model=mod_tobramycin_2cpt_fictional,
+                                  return_model = FALSE)
 
   expect_equal("data.frame" %in% attributes(p01_pop_mod$eta)$class,
                TRUE)
@@ -96,8 +96,11 @@ test_that("poso_simu_pop returns the expected objects", {
 })
 
 test_that("poso_estim_map returns the expected objects", {
-  p01_map_mod    <- poso_estim_map(patient01_tobra)
-  p01_map_nomod  <- poso_estim_map(patient01_tobra,return_model = FALSE)
+  p01_map_mod    <- poso_estim_map(dat=df_patient01_tobra,
+                                   prior_model=mod_tobramycin_2cpt_fictional)
+  p01_map_nomod  <- poso_estim_map(dat=df_patient01_tobra,
+                                   prior_model=mod_tobramycin_2cpt_fictional
+                                   ,return_model = FALSE)
 
   expect_equal("data.frame" %in% attributes(p01_map_mod$eta)$class,
                FALSE)
@@ -114,8 +117,12 @@ test_that("poso_estim_map returns the expected objects", {
 })
 
 test_that("poso_estim_mcmc returns the expected objects", {
-  p01_mcmc_mod   <- poso_estim_mcmc(patient01_tobra,burn_in=0,n_iter = 10)
-  p01_mcmc_nomod <- poso_estim_mcmc(patient01_tobra,burn_in=0,n_iter = 10,
+  p01_mcmc_mod   <- poso_estim_mcmc(dat=df_patient01_tobra,
+                                    prior_model=mod_tobramycin_2cpt_fictional,
+                                    burn_in=0,n_iter = 10)
+  p01_mcmc_nomod <- poso_estim_mcmc(dat=df_patient01_tobra,
+                                    prior_model=mod_tobramycin_2cpt_fictional,
+                                    burn_in=0,n_iter = 10,
                                     return_model = FALSE)
 
   expect_equal("data.frame" %in% attributes(p01_mcmc_mod$eta)$class,
@@ -129,10 +136,12 @@ test_that("poso_estim_mcmc returns the expected objects", {
 })
 
 test_that("poso_estim_sir returns the expected objects", {
-  p01_sir_mod   <- poso_estim_sir(patient01_tobra,
+  p01_sir_mod   <- poso_estim_sir(dat=df_patient01_tobra,
+                                  prior_model=mod_tobramycin_2cpt_fictional,
                                   n_sample=5e2,
                                   n_resample=1e2)
-  p01_sir_nomod <- poso_estim_sir(patient01_tobra,
+  p01_sir_nomod <- poso_estim_sir(dat=df_patient01_tobra,
+                                  prior_model=mod_tobramycin_2cpt_fictional,
                                   n_sample=5e2,
                                   n_resample=1e2,
                                   return_model = FALSE)
@@ -162,12 +171,12 @@ df_patient06_pipera <- data.frame(ID=6,TIME=c(0.0,8,16,24,32,40,47.9,48,
                                   OCC=c(1,1,1,1,1,1,1,2,2,2,2,3,3,3,3,
                                         4,4,4,4))
 
-patient06_pipera <- posologyr(prior_model=mod_piperacillin_2cpt_Roberts2010,
-                              dat=df_patient06_pipera)
-
 test_that("poso_simu_pop returns the expected objects (with IOV)", {
-  p06_pop_mod    <- poso_simu_pop(patient06_pipera)
-  p06_pop_nomod  <- poso_simu_pop(patient06_pipera,return_model = FALSE)
+  p06_pop_mod    <- poso_simu_pop(dat=df_patient06_pipera,
+                                  prior_model=mod_piperacillin_2cpt_Roberts2010)
+  p06_pop_nomod  <- poso_simu_pop(dat=df_patient06_pipera,
+                                  prior_model=mod_piperacillin_2cpt_Roberts2010,
+                                  return_model = FALSE)
 
   expect_equal("data.frame" %in% attributes(p06_pop_mod$eta)$class,
                TRUE)
@@ -180,8 +189,11 @@ test_that("poso_simu_pop returns the expected objects (with IOV)", {
 })
 
 test_that("poso_estim_map returns the expected objects (with IOV)", {
-  p06_map_mod    <- poso_estim_map(patient06_pipera)
-  p06_map_nomod  <- poso_estim_map(patient06_pipera,return_model = FALSE)
+  p06_map_mod    <- poso_estim_map(dat=df_patient06_pipera,
+                                   prior_model=mod_piperacillin_2cpt_Roberts2010)
+  p06_map_nomod  <- poso_estim_map(dat=df_patient06_pipera,
+                                   prior_model=mod_piperacillin_2cpt_Roberts2010,
+                                   return_model = FALSE)
 
   expect_equal("data.frame" %in% attributes(p06_map_mod$eta)$class,
                FALSE)
@@ -198,10 +210,12 @@ test_that("poso_estim_map returns the expected objects (with IOV)", {
 })
 
 test_that("poso_estim_sir returns the expected objects (with IOV)", {
-  p06_sir_mod   <- poso_estim_sir(patient06_pipera,
+  p06_sir_mod   <- poso_estim_sir(dat=df_patient06_pipera,
+                                  prior_model=mod_piperacillin_2cpt_Roberts2010,
                                   n_sample=5e2,
                                   n_resample=1e2)
-  p06_sir_nomod <- poso_estim_sir(patient06_pipera,
+  p06_sir_nomod <- poso_estim_sir(dat=df_patient06_pipera,
+                                  prior_model=mod_piperacillin_2cpt_Roberts2010,
                                   n_sample=5e2,
                                   n_resample=1e2,
                                   return_model = FALSE)