+++ tests/testthat/test_map_iov.R
@@ -59,19 +59,17 @@ df_patient06_pipera <- data.frame(ID=6,
                                   OCC=c(1,1,1,1,1,1,1,2,2,2,2,3,3,3,3,
                                        4,4,4,4))
 
-patient06_pipera <- posologyr(prior_model=mod_piperacillin_2cpt_Roberts2010,
-                             dat=df_patient06_pipera)
-
-patient06_pipera_map_iov <- poso_estim_map(patient06_pipera,
-                                      return_model=TRUE)
+pat06_map_iov <- poso_estim_map(dat=df_patient06_pipera,
+                                prior_model=mod_piperacillin_2cpt_Roberts2010,
+                                return_model=TRUE)
 
 test_that("MAP estimates match Monolix MAP estimates", {
-  expect_equal(patient06_pipera_map_iov$model$LAGTIME[1], 0.07, tolerance=1e-3)
-  expect_equal(unique(patient06_pipera_map_iov$model$Cl)[1], 10.58, tolerance=1e-3)
-  expect_equal(unique(patient06_pipera_map_iov$model$Cl)[2], 24.0, tolerance=1e-3)
-  expect_equal(unique(patient06_pipera_map_iov$model$Cl)[3], 18.4, tolerance=1e-3)
-  expect_equal(unique(patient06_pipera_map_iov$model$Cl)[4], 26.3, tolerance=1e-3)
-  expect_equal(patient06_pipera_map_iov$model$V1[1], 7.2, tolerance=1e-3)
-  expect_equal(patient06_pipera_map_iov$model$Q[1], 52, tolerance=1e-3)
-  expect_equal(patient06_pipera_map_iov$model$V2[1], 17.8, tolerance=1e-3)
+  expect_equal(pat06_map_iov$model$LAGTIME[1], 0.07, tolerance=1e-3)
+  expect_equal(unique(pat06_map_iov$model$Cl)[1], 10.58, tolerance=1e-3)
+  expect_equal(unique(pat06_map_iov$model$Cl)[2], 24.0, tolerance=1e-3)
+  expect_equal(unique(pat06_map_iov$model$Cl)[3], 18.4, tolerance=1e-3)
+  expect_equal(unique(pat06_map_iov$model$Cl)[4], 26.3, tolerance=1e-3)
+  expect_equal(pat06_map_iov$model$V1[1], 7.2, tolerance=1e-3)
+  expect_equal(pat06_map_iov$model$Q[1], 52, tolerance=1e-3)
+  expect_equal(pat06_map_iov$model$V2[1], 17.8, tolerance=1e-3)
 })