+++ tests/testthat/test_map_nm.R
@@ -45,8 +45,10 @@ pat_302_1 <- data.frame(ID=1,
                         BW=c(77,113,92,132,126,128,41,56,68,45,93,126),
                         SEX=c(0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 1))
 set.seed(0)
-pat302_nm  <- posologyr(mod_run302,dat=pat_302_1,nocb=TRUE)
-pat302_nm_map  <- poso_estim_map(pat302_nm,return_model=TRUE)
+
+pat302_nm_map  <- poso_estim_map(dat=pat_302_1,
+                                 mod_run302,
+                                 nocb=TRUE,return_model=TRUE)
 
 test_that("MAP estimates match NONMEM posthoc estimates (time varying covariates)", {
   expect_equal(pat302_nm_map$eta[1], c(ETA_Cl=0.398845969), tolerance=1e-3)